Move query client to tori module
