+++ libs/CMakeLists.txt
@@ -2,7 +2,6 @@ add_subdirectory(amount)
 add_subdirectory(common)
 add_subdirectory(crypto)
 add_subdirectory(logger)
-add_subdirectory(torii_utils)
 add_subdirectory(generator)
 add_subdirectory(parser)
 add_subdirectory(validator)
